a million changes, login, reformat profile, and more code organization
