+++ src/components/DailyProblem/ResizeBar.css
