+++ src/components/DailyProblem/ResizeBar.jsx
