--- src/components/login.css
@@ -1,118 +1,0 @@
-.mainlogin {
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: calc(550px + 10vh);
-}
-
-.loginCard {
-    /* border: 2px solid #e48257;*/
-    background-color: #2e2828;
-    border-radius: 30px;
-    max-width: 500px;
-    width: 85%;
-    /* height: 550px; */
-    padding-bottom: 2.5rem;
-    
-}
-
-.cardWrapper {
-    width: calc(100% - 50px);
-    height: calc(100% - 50px);
-    display: flex;
-    flex-direction: column; 
-    align-items: center;
-}
-
-.loginCardWrapped {
-    margin-top: 17px;
-    display: flex;
-    align-items: center;
-    /* justify-content: center; */
-    flex-direction: column;
-    position: relative;
-}
-
-.tigerlogo {
-    height: calc(50px + 5%);
-    width: calc(50px + 5%);
-    margin-top: 5%;
-    margin-bottom: 3%;
-}
-
-.enterTitle {
-    font-size: 25px;
-}
-
-@media (max-width: 450px) {
-    .enterTitle {
-        font-size: 6vw;
-    }   
-
-    .enterForm {
-        margin-top: 3vw !important;
-        /* width: 80% !important; */
-    }
-}
-
-.enterForm {
-    display: flex;
-    flex-direction: column;
-    gap: 5px;
-    margin-top: 20px;
-    width: 80%;
-    max-width: 300px;
-}
-
-.enterInput {
-    width: 100%;
-    padding: 7px;
-    border-radius: 10px;
-    border: 2px solid #473f3f;
-    background-color: #393232;
-    color: #a19090 !important;
-}
-
-.enterInput:focus {
-    outline: none;
-    /* border: 2px solid #e48257; */
-}
-
-.enterInput::placeholder {
-    color: #a19090;
-}
-
-.enterSubmit {
-    margin-top: 5px;
-    border-radius: 10px;
-    height: 40px;
-    transition: background-color .3s ease;
-    font-weight: bold;
-}
-
-.enterActivated {
-    background-color: #c1501f !important;
-    color: white;
-    border: none;
-    cursor: pointer;
-}
-
-.enterDeactivated {
-    background-color: #231f1f;
-    color: #a0aec0;
-    border: none;
-    cursor: not-allowed !important;
-}
-
-.errMSG {
-    margin-top: 5px;
-    text-align: center;
-    font-size: 12px;
-}
-
-.switchEnter {
-    font-size: 12px;
-    text-align: center;
-    margin-top: 2px;
-}
