+++ src/pages/Login.jsx
@@ -1,0 +1,256 @@
+import React, { useState } from "react";
+import { Navigate, Link } from "react-router-dom";
+import {
+  doSignInWithEmailAndPassword,
+  doSignInWithGoogle,
+  doPasswordReset,
+} from "../firebase/auth";
+import { useAuth } from "../contexts/authContext";
+
+import "./RegisterLogin.css";
+
+const Login = () => {
+  const { userLoggedIn } = useAuth();
+
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [isSigningIn, setIsSigningIn] = useState(false);
+  const [errorMessage, setErrorMessage] = useState("");
+  const [resetMessage, setResetMessage] = useState("");
+
+  const onSubmit = async (e) => {
+    e.preventDefault();
+    if (!isSigningIn) {
+      setIsSigningIn(true);
+      setErrorMessage(""); // Clear any previous error messages
+
+      // Check email format
+      const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+      if (!emailRegex.test(email)) {
+        setErrorMessage("Please enter a valid email address");
+        setIsSigningIn(false);
+        return;
+      }
+
+      try {
+        await doSignInWithEmailAndPassword(email, password);
+        // doSendEmailVerification()
+      } catch (error) {
+        console.error("Login error:", error);
+        // Handle different types of Firebase auth errors
+        let errorMsg = "An error occurred during sign in";
+        if (error.code === "auth/user-not-found") {
+          errorMsg = "No account found with this email address";
+        } else if (error.code === "auth/wrong-password") {
+          errorMsg = "Incorrect password";
+        } else if (error.code === "auth/invalid-credential") {
+          errorMsg = "Incorrect email or password";
+        } else if (error.code === "auth/too-many-requests") {
+          errorMsg =
+            "Too many failed attempts. Please try again later";
+        } else if (error.code === "auth/user-disabled") {
+          errorMsg = "This account has been disabled";
+        }
+        setErrorMessage(errorMsg);
+      } finally {
+        setIsSigningIn(false);
+      }
+    }
+  };
+
+  const onGoogleSignIn = async (e) => {
+    e.preventDefault();
+    if (!isSigningIn) {
+      setIsSigningIn(true);
+      setErrorMessage(""); // Clear any previous error messages
+      try {
+        await doSignInWithGoogle();
+      } catch (error) {
+        console.error("Google sign-in error:", error);
+        let errorMsg = "An error occurred during Google sign in";
+        if (error.code === "auth/popup-closed-by-user") {
+          errorMsg = "Sign in was cancelled";
+        } else if (error.code === "auth/popup-blocked") {
+          errorMsg =
+            "Popup was blocked. Please allow popups for this site";
+        } else if (
+          error.code ===
+          "auth/account-exists-with-different-credential"
+        ) {
+          errorMsg =
+            "An account already exists with the same email address but different sign-in credentials";
+        }
+        setErrorMessage(errorMsg);
+      } finally {
+        setIsSigningIn(false);
+      }
+    }
+  };
+
+  const onForgotPassword = async () => {
+    if (!email) {
+      setErrorMessage("Please enter your email address first");
+      return;
+    }
+    setErrorMessage("");
+    setResetMessage("");
+    try {
+      await doPasswordReset(email);
+      setResetMessage(
+        "Password reset email sent! Check your inbox and junk mail."
+      );
+    } catch (error) {
+      console.error("Password reset error:", error);
+      setErrorMessage(
+        "Failed to send password reset email. Please try again."
+      );
+    }
+  };
+
+  return (
+    <div>
+      {userLoggedIn && <Navigate to={"/"} replace={true} />}
+
+      <main className="mainlogin">
+        <div className="loginCard">
+          <div className="loginCardWrapped">
+            <img
+              className="tigerlogo"
+              src="images/icons/tiger icon.png"
+              alt="Tiger Logo"
+            />
+            <div className="text-center">
+              <div className="mt-2">
+                <h3 className="enterTitle">Welcome Back</h3>
+              </div>
+            </div>
+            <form onSubmit={onSubmit} className="enterForm">
+              <div>
+                <input
+                  placeholder="Email"
+                  type="email"
+                  autoComplete="email"
+                  required
+                  value={email}
+                  onChange={(e) => {
+                    setEmail(e.target.value);
+                  }}
+                  className="enterInput"
+                />
+              </div>
+
+              <div>
+                <input
+                  placeholder="Password"
+                  type="password"
+                  autoComplete="current-password"
+                  required
+                  value={password}
+                  onChange={(e) => {
+                    setPassword(e.target.value);
+                  }}
+                  className="enterInput"
+                />
+              </div>
+
+              <div className="text-right">
+                <button
+                  type="button"
+                  onClick={onForgotPassword}
+                  className="text-sm text-blue-600 hover:underline"
+                >
+                  Forgot Password?
+                </button>
+              </div>
+
+              {errorMessage && (
+                <span className="errMSG">{errorMessage}</span>
+              )}
+
+              {resetMessage && (
+                <span className="text-green-600 text-sm">
+                  {resetMessage}
+                </span>
+              )}
+
+              <button
+                type="submit"
+                disabled={isSigningIn}
+                className={`enterSubmit ${password && email
+                    ? "enterActivated"
+                    : "enterDeactivated"
+                  }`}
+              >
+                {isSigningIn ? "Signing In..." : "Sign In"}
+              </button>
+            </form>
+            <p className="switchEnter">
+              Don't have an account?{" "}
+              <Link
+                to={"/register"}
+                className="hover:underline font-bold"
+              >
+                Sign up
+              </Link>
+            </p>
+            <div className="flex flex-row text-center w-full">
+              <div className="border-b-2 mb-2.5 mr-2 w-full"></div>
+              <div className="text-sm font-bold w-fit">OR</div>
+              <div className="border-b-2 mb-2.5 ml-2 w-full"></div>
+            </div>
+            <button
+              disabled={isSigningIn}
+              onClick={(e) => {
+                onGoogleSignIn(e);
+              }}
+              className={`w-full flex items-center justify-center gap-x-3 py-2.5 border rounded-lg text-sm font-medium  ${isSigningIn
+                  ? "cursor-not-allowed"
+                  : "hover:bg-gray-100 transition duration-300 active:bg-gray-100"
+                }`}
+            >
+              <svg
+                className="w-5 h-5"
+                viewBox="0 0 48 48"
+                fill="none"
+                xmlns="http://www.w3.org/2000/svg"
+              >
+                <g clipPath="url(#clip0_17_40)">
+                  <path
+                    d="M47.532 24.5528C47.532 22.9214 47.3997 21.2811 47.1175 19.6761H24.48V28.9181H37.4434C36.9055 31.8988 35.177 34.5356 32.6461 36.2111V42.2078H40.3801C44.9217 38.0278 47.532 31.8547 47.532 24.5528Z"
+                    fill="#4285F4"
+                  />
+                  <path
+                    d="M24.48 48.0016C30.9529 48.0016 36.4116 45.8764 40.3888 42.2078L32.6549 36.2111C30.5031 37.675 27.7252 38.5039 24.4888 38.5039C18.2275 38.5039 12.9187 34.2798 11.0139 28.6006H3.03296V34.7825C7.10718 42.8868 15.4056 48.0016 24.48 48.0016Z"
+                    fill="#34A853"
+                  />
+                  <path
+                    d="M11.0051 28.6006C9.99973 25.6199 9.99973 22.3922 11.0051 19.4115V13.2296H3.03298C-0.371021 20.0112 -0.371021 28.0009 3.03298 34.7825L11.005 19.4115C12.901 13.7235 18.2187 9.49932 24.48 9.49932Z"
+                    fill="#FBBC04"
+                  />
+                  <path
+                    d="M24.48 9.49932C27.9016 9.44641 31.2086 10.7339 33.6866 13.0973L40.5387 6.24523C36.2 2.17101 30.4414 -0.068932 24.48 0.00161733C15.4055 0.00161733 7.10718 5.11644 3.03296 13.2296L11.005 19.4115C12.901 13.7235 18.2187 9.49932 24.48 9.49932Z"
+                    fill="#EA4335"
+                  />
+                </g>
+                <defs>
+                  <clipPath id="clip0_17_40">
+                    <rect
+                      width="48"
+                      height="48"
+                      fill="white"
+                    />
+                  </clipPath>
+                </defs>
+              </svg>
+              {isSigningIn
+                ? "Signing In..."
+                : "Continue with Google"}
+            </button>
+          </div>
+        </div>
+      </main>
+    </div>
+  );
+};
+
+export default Login;
